implemented parsing and executing of task
mgmt is our entry point. It will require at the least a yaml configuration
file. It will be used by the parse to figure out what kind of action we
want to execute on the server.

The parser after validating the yaml syntax will try to match the action with
the action specified in the operations_registry.

If the action has not been defined already we can define it in code via the
approriate API (ops_registry.py)

If an action is matched the tool will try to establish if the configuration has
all the mandatory arguments specified. If so with also the optional one possibly
specified it will run the defined action(s).

This is possible through the use of inspect and functool python module in the
standard library.

There is a possible attack vector, although I wasnt able to exploit, by somehow
mangling the possible positional argument via the use of functools.partial.