introduce basic depencendy handling
Dependency handling is done by looking at a new meta->dependencies dictionary
in the action definition. Unfortunately we support only one level of depedency. Dependencies
of a dependency wont be executed until the dependency is solely executed. Its not nice but
for the scope of the exercise I think its enough.